third_party/blink/web_tests/wpt_internal/geolocation-api/watchPosition-page-visibility.https.html

<!DOCTYPE html>
<html>
<head>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-adapter.js"></script>
</head>
<body>
<script type="module">
import {GeolocationMock} from './resources/geolocation-mock.js';

description("Tests that watchPosition does not report position changes when the page is not visible.");

let position;
let error;
let isPageVisible = true;

const mock = new GeolocationMock();
mock.setGeolocationPermission(true);

let mockLatitude = 51.478;
let mockLongitude = -0.166;
const mockAccuracy = 100.0;

function updatePosition() {
    ++mockLatitude;
    ++mockLongitude;
    mock.setGeolocationPosition(mockLatitude, mockLongitude, mockAccuracy);
}

function setMainWindowHidden(hidden) {
  return new Promise(resolve => {
    document.addEventListener('visibilitychange', resolve, {once: true});
    testRunner.setMainWindowHidden(hidden);
  });
}

function roundTripToBrowser() {
  return new Promise(resolve => {
    // Use an OOPIF navigation to elicit a browser round-trip.
    const frame = document.createElement('iframe');
    frame.src = 'http://localhost:8080/resources/blank.html';
    frame.addEventListener('load', () => {
      document.body.removeChild(frame);
      resolve();
    });
    document.body.appendChild(frame);
  });
}

updatePosition();

let state = 0;

function checkPosition(p) {
    position = p;
    assert_equals(position.coords.latitude, mockLatitude);
    assert_equals(position.coords.longitude, mockLongitude);
}

async function showPageAndUpdatePosition() {
    assert_false(isPageVisible);
    state++;
    await setMainWindowHidden(false);
    isPageVisible = true;
    updatePosition();
}

navigator.geolocation.watchPosition(async p => {
    assert_true(isPageVisible);
    state++;
    checkPosition(p);
    switch(state) {
        case 1:
            updatePosition();
            break;
        case 2: {
            await setMainWindowHidden(true);
            isPageVisible = false;
            // This should not be received.
            updatePosition();
            // Wait for it to not be received, then continue.
            await roundTripToBrowser();
            showPageAndUpdatePosition();
            break;
        }
        case 3:
            updatePosition();
            break;
        case 4:
            finishJSTest();
            return;
    }
}, function(e) {
    testFailed('Error callback invoked unexpectedly');
    finishJSTest();
});

</script>
</body>
</html>
